Refactoring
- Started refactoring units to easily implement new patterns
