+++ Assets/Scripts/Units/Attackers/Attack_Melee_All.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Attack_All : MonoBehaviour
+public class Attack_Melee_All : MonoBehaviour
 {
     private void OnTriggerEnter2D(Collider2D collision)
     {
